Add wrapper around unordered_map so hash<value>==key
Make this wrapper so instead of having to do the logic every time
over this wrapper does it for you. I don't understand why there
isn't a STL library for this, but in my basic tests the speeds
are the same (within variation, maybe 100ms on 6seconds on 1M
items and several thousands of lookups) and the logic also works
correctly.